chore: renaming component files
