+++ src/components/Breadcrumb/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Icon from '../Icon';
+import Icon from '../Icon/Icon';
 
 interface BreadcrumbProps {
   isFinal?: boolean;
