+++ src/components/MessageInput/MessageInput.tsx
@@ -1,7 +1,7 @@
 import React, { useLayoutEffect, useRef } from 'react';
 
 import { SendButton } from '../ConversationStarter/SendButton';
-import Icon from '../Icon';
+import Icon from '../Icon/Icon';
 
 interface MessageInputProps {
   onChange: (value: string) => void;
