+++ stories/Alert.stories.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Alert from '../src/components/Alert';
+import Alert from '../src/components/Alert/Alert';
 
 export default {
   title: 'Alert',
